Add additional annotations to private/topology/underlay
In particular, add postconditions asserting that results are low
wherever it is applicable, in order to make the contracts more precise.
This is in anticipation of needing this information for further
verification.

Note that I still have some concerns regarding the pre- and
postconditions of `fmt.Sprintf` and `serrors.New` which are elaborated
in comments.